git-svn-id: svn://123.57.173.36/meijia-parent@1221 52ac4245-7803-4e87-b018-04b7c5ce731e

+++ simi-dao/src/main/java/com/simi/po/dao/user/UserLeaveMapper.java
@@ -8,9 +8,9 @@ import com.simi.vo.UserLeaveSearchVo;
 public interface UserLeaveMapper {
     int deleteByPrimaryKey(Long leaveId);
 
-    int insert(UserLeave record);
+    Long insert(UserLeave record);
 
-    int insertSelective(UserLeave record);
+    Long insertSelective(UserLeave record);
 
     UserLeave selectByPrimaryKey(Long leaveId);
